% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_cl}
\alias{l_cl}
\title{Table of the structure of phytosociological classes}
\usage{
l_cl(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table of the structure of classes in flora. The 10 most abundant are plotted, and rest of them are grouped as "other"
It may be also used to plotting another classifications - this function is flexible
}

