% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_csr}
\alias{l_csr}
\title{Table of the structure of Grime's life strategies}
\usage{
l_csr(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table of the structure of Grime's life strategies in flora.
The values may be taken from BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)
}

