% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_fam}
\alias{l_fam}
\title{Table of the structure of families}
\usage{
l_fam(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table representing structure of families in flora.
The share 10 most abundant families are calculated, and rest of them are grouped as "other".
}

