% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_hem}
\alias{l_hem}
\title{Table of the structure of hemeroby levels}
\usage{
l_hem(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table representing structure of hemeroby levels in flora.
The share of all hhemeroby levels are calculated. The usage of this index is constrained only to the Poznań, however this function may be useful for anyone who will want to modify it for own usage.
}
\references{
Jackowiak, B. 1993. Atlas rozmieszczenia roślin naczyniowych w Poznaniu (Atlas of distribution of vascular plants in Poznań). Prace Zakładu Taksonomii Roślin UAM w Poznaniu nr 2 (Publications of the Department of Plant Taxonomy of the Adam Mickiewicz University in Poznań – No 2). 409 pp. (Polish/English text).
}

