% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_hg}
\alias{l_hg}
\title{Table of the structure of historical-geographical groups}
\usage{
l_hg(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table representing structure of historical-geographical groups in flora.
The share of all historical-geographical groups are calculated
}

