% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_l}
\alias{l_l}
\title{Table of the structure of Ellenberg's ligth indicator}
\usage{
l_l(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table of the structure of Ellenberg's ligth indicator in flora.
1 represents shadow-demaning species, 9 - light-demanding species.
The values may be taken from Ellenberg and Leuschner (2010)
}
\references{
Ellenberg H., Leuschner C. 2010. Vegetation Mitteleuropas mit den Alpen in ökologischer, dynamischer und historischer Sicht. UTB UmgH.
}

