% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_lfo}
\alias{l_lfo}
\title{Table of the structure of life forms}
\usage{
l_lfo(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function retutns the table of the structure of life forms in flora.
Nanophanerophytes and megaphanerophytes were grouped as Phanerophytes (F), woody chamaephytes  and herbaceus chamaephytes were grouped as chamaephytes (C)
T - terophytes, Hy - hydrophytes, G - geophytes, H - hemicryptophytes
}

