% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_m}
\alias{l_m}
\title{Table of the structure of Ellenberg's moisture indicator}
\usage{
l_m(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of Ellenberg's moisture indicator in flora.
1 represents species of dry soils, 5 - moderatly moisture soils, 10 - wet soils and 12 - aquatic species (e.g. Hottonia palustris).
The values may be taken from Ellenberg and Leuschner (2010)
}
\references{
Ellenberg H., Leuschner C. 2010. Vegetation Mitteleuropas mit den Alpen in ökologischer, dynamischer und historischer Sicht. UTB UmgH.
}

