% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_sr}
\alias{l_sr}
\title{Table of the structure of Ellenberg's soil reaction indicator}
\usage{
l_sr(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table of the structure of Ellenberg's soil reaction indicator in flora.
1 represents strongly acidophilous species, 9 - basiphilous species.
The values may be taken from Ellenberg and Leuschner (2010)
}
\references{
Ellenberg H., Leuschner C. 2010. Vegetation Mitteleuropas mit den Alpen in ökologischer, dynamischer und historischer Sicht. UTB UmgH.
}

