% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l_urb}
\alias{l_urb}
\title{Table of the structure of urbanity}
\usage{
l_urb(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table of the structure of urbanity in flora.
1 represents urbanophobus species, 3 - urbanoneutral, and 5 - urbanophilous.
The values may be taken from BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)
}

