\name{MKDgeobot-package}
\alias{MKDgeobot-package}
\alias{MKDgeobot}
\docType{package}
\title{
Tools for geobotanical analysis
}
\description{
Functions for flora and vegetation analysis. Extracting ecological indices and ploting them using ggplot2.
}
\details{
\tabular{ll}{
Package: \tab MKDgeobot\cr
Type: \tab Package\cr
Version: \tab 0.3.0.2\cr
Date: \tab 2014-12-07\cr
License: \tab GPL 2.0\cr
}
This package contains very simple functions to group and plot share of ecological groups in floristic list and phytosociological releves. Due to local research custosm, some of these functions may be not usefull. Some of them may require modification to aim your goals. In package I also contained example datasets: one with phytosociological releves and one with floristic list. The usage of functions need `data frame` with colnames like in example. It is (unfortunetaly) essential to **keep the order and names of columns** - because function `tabelka()` is very simple/primitive. 
}
\author{
Marcin K. Dyderski
Complains, conclusions, improvements: <Marcin.Dyderski@gmail.com>
}
\references{
Jackowiak, B. 1993.Atlas rozmieszczenia roślin naczyniowych w Poznaniu (Atlas of distribution of vascular plants in Poznań). Prace Zakładu Taksonomii Roślin UAM w Poznaniu nr 2 (Publications of the Department of Plant Taxonomy of the Adam Mickiewicz University in Poznań – No 2). 409 pp. (Polish/English text).

Ratyńska H., Wojterska M., Brzeg A., Kołacz M., 2011. Multimedialna encyklopedia zbiorowisk roślinnych Polski. NFOSiGW, UKW, IETI.

Dzwonko Z., Loster S. 2001. Wskaźnikowe gatunki starych lasów i ich znaczenie dla ochrony przyrody i kartografii roślinności. IGiPZ PAN, Prace Geogr., 178: 120–132.

Ellenberg H., Leuschner C. 2010. Vegetation Mitteleuropas mit den Alpen in ökologischer, dynamischer und historischer Sicht. UTB UmgH.
}

\keyword{ geobotany, flora, vegetation }

