% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pceco}
\alias{pceco}
\title{Data about 130 species of plants}
\format{A data frame with 130 rows and 21 variables
\describe{
\item{name}{Latin name of species, without authors}
\item{fam}{Latin name of the family}
\item{hg}{Historical-ecological group, after Jackowiak (1993): sp-spontaneophytes, ap-apophytes, arch-archaeophytes, kn-kenophytes, ef-ephemerophytes}
\item{hemero}{Level of hemeroby, after Jackowiak (1993): 0 - species without synanthropodynamic tendences, 1-6: hemerophilous species (the higher value, the level of fidelity to transformed habitats higher), characters: hemerophobus species (fidelity to non-transformed habitats) - threat categories: PR - potentially vulnerable, I- not certain, p1-p3 - potentially endangered, V- vulnerable, E- endangered, Ex - extinct}
\item{soceko}{Scoiological-ecological groups after Jackowiak (1993): not usefull outside Poznań, prepared for author's convience}
\item{lf}{Life forms according to Raunkiaer's classification, after Jackowiak (1993): M- megaphanerophytes, N-nanophanerophytes, C- chamaephytes (woody), Ch- chamaephytes (herbaceus), T - terophytes, Hy - hydrophytes, G - geophytes, H - hemicryptophytes}
\item{lf_opis}{descriptive, non-analytical list of ALL potential lifeforms}
\item{class}{Phytosociological class, after Ratyńska et al. (2011)}
\item{choria}{NON analytic- decriptive, all potential modes of dispersal, after BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)}
\item{chory}{Main mode of dispersal, after BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)}
\item{pollen.vector}{NON analytic- decriptive, all potential polination vectors, after BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)}
\item{csr}{Type of Grime's life strategy, after BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)}
\item{urbanity}{level of urbanity in scale from 1 - urbanophobous to 5 - urbanophilous, after BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)}
\item{stare.lasy}{Is species a Ancient Woodland Indicator species? 0 - no, 1- yes. After Dzownko and Loster (2001)}
\item{L}{Ellenberg's ligth indicator, after Ellenberg and Leuschner (2010)}
\item{T}{Ellenberg's temperature indicator, after Ellenberg and Leuschner (2010)}
\item{C}{Ellenberg's continentality indicator, after Ellenberg and Leuschner (2010)}
\item{M}{Ellenberg's moisture indicator, after Ellenberg and Leuschner (2010)}
\item{SR}{Ellenberg's soil reaction indicator, after Ellenberg and Leuschner (2010)}
\item{N}{Ellenberg's fertility indicator, after Ellenberg and Leuschner (2010)}
}}
\description{
This dataset contains information about 130 species of vascular plants and mosses.
}
\references{
Jackowiak, B. 1993.Atlas rozmieszczenia roślin naczyniowych w Poznaniu (Atlas of distribution of vascular plants in Poznań). Prace Zakładu Taksonomii Roślin UAM w Poznaniu nr 2 (Publications of the Department of Plant Taxonomy of the Adam Mickiewicz University in Poznań – No 2). 409 pp. (Polish/English text).

Ratyńska H., Wojterska M., Brzeg A., Kołacz M., 2011. Multimedialna encyklopedia zbiorowisk roślinnych Polski. NFOSiGW, UKW, IETI.

Dzwonko Z., Loster S. 2001. Wskaźnikowe gatunki starych lasów i ich znaczenie dla ochrony przyrody i kartografii roślinności. IGiPZ PAN, Prace Geogr., 178: 120–132.

Ellenberg H., Leuschner C. 2010. Vegetation Mitteleuropas mit den Alpen in ökologischer, dynamischer und historischer Sicht. UTB UmgH.
}

