% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pctab}
\alias{pctab}
\title{Phytosociological releves from Poznań}
\format{A data frame with 25 rows and 170 variables}
\source{
releves were send to Polish Vegetation Database \url{http://synbiot.uni.wroc.pl/en/pages/1000.html}
}
\description{
This dataset contains 25 releves from Poznań.
The releves were conducted from 2012-2014 in 5 types of forest plant communities: alder carrs Carici elongatae-Alnetum (rel. 1-5), ash-alder riparian forests Fraxino-Alnetum (rel. 6-10), Scots pine plantations - community Pinus-Padus (rel. 11-15), Black locust spontaneous tree stands (Chelidonio-Robinietum) and in oak-hornbeam forests (Galio-Carpinetum).
The names of columns are names of species, with number of vegetation layer: 1 - tree layer 2 - lower tree layer (a2), 4- shrub layer, 6 - herb layer, 9 - moss layer
}
\details{
The data was collected by Marcin K. Dyderski during geobotanical investigation of urban forests in Poznań (papers are in progres and in print, when will be avaiable, references will be set up). Data was deposed in Polish Vegetation Database.
}

