% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabelka}
\alias{tabelka}
\title{Phytosociological releves analysis}
\usage{
tabelka(eco, tab)
}
\arguments{
\item{eco}{Tabelka ze wskaźnikami ekologicznymi (baza danych)}

\item{tab}{tabela fitosocjologiczna}
}
\description{
This function provides tools to extract synthetic ecological variables from releves. It is constrained to Polish specific conditions and research traditions. However, it supply the quantitative analysis of material and let it be analysed.
I don't know what do you need, so feel free to modify the body of the function by adding your custom indices.
Mean Ellenberg's indicator values are means weighted by species' abundance.
}

