% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_chory}
\alias{wyk_chory}
\title{Plot the structure of ways of dispersal}
\usage{
wyk_chory(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of ways of dispersal in flora.
The values may be taken from BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)
}

