% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_csr}
\alias{wyk_csr}
\title{Plot the structure of Grime's life strategies}
\usage{
wyk_csr(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of Grime's life strategies in flora.
The values may be taken from BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)
}

