% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_f}
\alias{wyk_f}
\title{Plot the structure of Ellenberg's soil fertility indicator}
\usage{
wyk_f(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function returns the table of the structure of Ellenberg's soil fertility indicator in flora.
1 represents strongly poor-nutrient soil species, 9 - strongly nutrient-rich soil species.
The values may be taken from Ellenberg and Leuschner (2010)
}
\references{
Ellenberg H., Leuschner C. 2010. Vegetation Mitteleuropas mit den Alpen in ökologischer, dynamischer und historischer Sicht. UTB UmgH.
}

