% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_fam}
\alias{wyk_fam}
\title{Plot the structure of families}
\usage{
wyk_fam(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of families in flora. The 10 most abundant families are plotted, and rest of them are grouped as "other"
}

