% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_hg}
\alias{wyk_hg}
\title{Plot the structure of historical-geographical groups}
\usage{
wyk_hg(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of historical-geographical groups in flora.
The groups are plotted in order of anthropopressure responses: spontaneophytes, apophytes, archaeophytes, kenophytes and ephemerophytes
}

