% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_lfo}
\alias{wyk_lfo}
\title{Plot the structure of life forms}
\usage{
wyk_lfo(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of life forms in flora.
Nanophanerophytes and megaphanerophytes were grouped as Phanerophytes (F), woody chamaephytes  and herbaceus chamaephytes were grouped as chamaephytes (C)
T - terophytes, Hy - hydrophytes, G - geophytes, H - hemicryptophytes
}

