% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_pollen}
\alias{wyk_pollen}
\title{Plot the structure of mode of polinations}
\usage{
wyk_pollen(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of mode of polinations in flora.
The values may be taken from BiolFlor database (http://www2.ufz.de/biolflor/index.jsp)
}

