% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyk_soceko}
\alias{wyk_soceko}
\title{Plot the structure of sociological-ecological groups}
\usage{
wyk_soceko(co)
}
\arguments{
\item{co}{"eco" data.frame, containing flora and species' properties}
}
\description{
This function plots the structure of sociological-ecological groups in flora.
'The usage of this index is constrained only to flora of Poznań, however it may be usefull for someone to modify this function.
Sociological-ecological groups  are taken by Jackowiak (1993) and shows ecological scales
For better representation of sociological-ecological properties in other areas look for MKDmisc:::wyk_cl
}
\references{
Jackowiak, B. 1993. Atlas rozmieszczenia roślin naczyniowych w Poznaniu (Atlas of distribution of vascular plants in Poznań). Prace Zakładu Taksonomii Roślin UAM w Poznaniu nr 2 (Publications of the Department of Plant Taxonomy of the Adam Mickiewicz University in Poznań – No 2). 409 pp. (Polish/English text).
#'
}

