\name{MKDgeobot-package}
\alias{MKDgeobot-package}
\alias{MKDgeobot}
\docType{package}
\title{
Tools for geobotanical analysis
}
\description{
Functions for flora and vegetation analysis. Extracting ecological indices and ploting them using ggplot2.
}
\details{
\tabular{ll}{
Package: \tab MKDgeobot\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2014-12-07\cr
License: \tab GPL 2.0\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Marcin K. Dyderski
Complains, conclusions, improvements: <Marcin.Dyderski@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
